Configure STRAPS bits for hard-coded GPU device.
Use hard-coded location of the GPU on the PCIe bus, as well as the
associated PCI-to-PCI bridge, to map BAR0 during PCIe enumeration,
and configure STRAPS bits to set larger size for BAR1.

BAR0 is temporarily mapped to the hard-coded memory and I/O range
addresses (manually read from GPU-Z report), so the same values can be
used as for the running (already configured) system.