Add change detection for Setup variable.
Compute and save the CRC64 of the Setup variable, or the Custom variable
if Setup is missing, and use the saved CRC64 to detect changes in UEFI
settings.

Instances of the Setup variable that are less then 16 bytes in length
are ignored.

When a change is detected the DXE driver configuration is cleared (the
driver is disabled).

Add user option to disable the change detection.